% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{apple_captive_test}
\alias{apple_captive_test}
\title{Download Apple's captive portal test}
\usage{
apple_captive_test()
}
\description{
If the test page, returns "Success" that means that the computer is
connected to the Internet.
}
\details{
Note that this function will fail if the computer is offline. Use
\code{\link[=is_online]{is_online()}} to check if the computer is online.
}
\examples{
apple_captive_test()
}
