% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-ip.R
\name{my_ip}
\alias{my_ip}
\title{Query the computer's public IP address}
\usage{
my_ip(method = c("dns", "https"))
}
\arguments{
\item{method}{Whether to use a DNS or HTTPS query.}
}
\value{
Computer's public IP address as a string.
}
\description{
It can use a DNS query to opendns.com, if \code{method == "dns"}, or
an HTTPS query to icanhazip.com, see https://github.com/major/icanhaz.
The DNS query is much faster, the HTTPS query is secure.
}
\examples{
my_ip()
my_ip(method = "https")
}
