% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping-package.r
\name{ping}
\alias{ping}
\title{Ping a remote server, to see if it is alive}
\usage{
ping(
  destination,
  continuous = FALSE,
  verbose = continuous,
  count = 3L,
  timeout = 1
)
}
\arguments{
\item{destination}{Host name or IP address.}

\item{continuous}{Logical, whether to keep pinging until the
user interrupts.}

\item{verbose}{Whether to print progress on the screen while
pinging.}

\item{count}{Number of pings to perform.}

\item{timeout}{Timeout for a ping response.}
}
\value{
Vector of response times. \code{NA} means no response, in
milliseconds. Currently \code{NA}s are always at the end of the vector,
and not in their correct position.
}
\description{
This is the classic ping, using ICMP packages. Only the
system administrator can send ICMP packages, so we call out
to the system's ping utility.
}
\examples{
ping("8.8.8.8")
ping("r-project.org")
}
