% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ping_port}
\alias{ping_port}
\title{Check if a port of a server is active, measure response time}
\usage{
ping_port(destination, port = 80L, type = c("tcp", "udp"),
  continuous = FALSE, count = 3L, timeout = 1)
}
\arguments{
\item{destination}{Host name or IP address.}

\item{port}{Port.}

\item{type}{Protocol, \sQuote{tcp} or \sQuote{udp}.}

\item{continuous}{Logical, whether to keep pinging until
the user interrupts.}

\item{count}{Number of pings to perform.}

\item{timeout}{Timeout, in seconds. How long to wait for a
ping to succeed.}
}
\value{
Vector of response times. \code{NA} means no response
}
\description{
Check if a port of a server is active, measure response time
}

