% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-stop-pinging.r
\name{startPinging}
\alias{startPinging}
\title{Turns on notification sound for finished code}
\usage{
startPinging(wait = 30, ping = 2)
}
\arguments{
\item{wait}{In seconds, the minimum run time before a notification sound is made.}

\item{ping}{A number between 1-9 choosing the notification sound (see \link[beepr]{beep} for choices)}
}
\description{
Turns on notification sound for finished code
}
\examples{
\dontrun{
startPinging(wait = 30, ping = 5)
}
}
