% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping-package.r
\name{is_online}
\alias{is_online}
\title{Is the computer online?}
\usage{
is_online(timeout = 0.2)
}
\arguments{
\item{timeout}{Timeout for the pings.}
}
\value{
Possible values: \itemize{
\item \code{TRUE} Yes, online.
\item \code{FALSE} No, not online.
\item \code{"nodns"} We re online, but without a DNS service.
}
}
\description{
Ping some name servers that are always (well, almost) up.
If these are unreachable, then you are most probably not online.
}
\examples{
is_online()
is_online(timeout = 0.01)
}
