% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping-package.R, R/pingr-package.R
\docType{package}
\name{pingr-package}
\alias{pingr}
\alias{pingr-package}
\title{Check if the local or remote computer is up}
\description{
Check if a remote computer is up. It can either just call the system ping command, or check a specified TCP port.

Check if a remote computer is up. It can either just call the system ping command, or check a specified TCP port.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-lib.github.io/pingr/}
  \item \url{https://github.com/r-lib/pingr}
  \item Report bugs at \url{https://github.com/r-lib/pingr/issues}
}


Useful links:
\itemize{
  \item \url{https://r-lib.github.io/pingr/}
  \item \url{https://github.com/r-lib/pingr}
  \item Report bugs at \url{https://github.com/r-lib/pingr/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
